Make .keys() return a tuple with keys in sorted order (this last thing comes free from the skiplist implementation)

+++ skip.h
@@ -15,7 +15,7 @@ typedef struct _skipitem {
 typedef struct {
   PyObject_HEAD
   unsigned int level;
-  unsigned int items_used;
+  Py_ssize_t items_used;
   skipitem *header;
   PyTypeObject *type;
 } SkipDict;
